BACKEND/FastAPI/requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
crepe
scipy
tqdm
pandas
mediapipe
scikit-learn
librosa
tensorflow
tensorflow-hub
pyaudio
fastapi
uvicorn
websockets
python-multipart
pydantic
ultralytics
PyYAML
black
